+++ include/sdf/JointAxis.hh
@@ -207,6 +207,12 @@ namespace sdf
     /// \sa double Dissipation() const
     public: void SetDissipation(const double _dissipation) const;
 
+    /// \brief Get a pointer to the SDF element that was used during
+    /// load.
+    /// \return SDF element pointer. The value will be nullptr if Load has
+    /// not been called.
+    public: sdf::ElementPtr Element() const;
+
     /// \brief Private data pointer
     private: JointAxisPrivate *dataPtr;
   };